copy object optimization: async on AWS iff owners are same
If, when calling copyObject(key,other bucket) on a Bucket instnace, the
client's account owner IDs are the same, then we will kick off a copy
job on the AWS backend asynchronously. This will not incurr any network
costs on our side.

Otherwise, we will resort to copying the object contents directly to
this JVM and then send the contents back to AWS.